rc: mount virtual fs and /run before starting udevd
This commit is part of splitting pkgsrc-core/eudev/start_udev script by
moving /dev & /run mounting to rc.

So, if /dev is not mounted, mount as a devtmpfs (CONFIG_DEVTMPFS=y).
Otherwise, if /dev is mounted (e.g. due to handover from initramfs or
CONFIG_DEVTMPFS_MOUNT=y), remount with specific options:

  * some video drivers require "exec" access in /dev, thus it's set here
  * for completeness, add some sanity limits (2k non-empty files, 16k
    inodes)

The /run mountpoint is required by udevd, which we're calling later.